requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic-settings
python-multipart
openai

